AO3: fix fic-identity hash mismatch + finish X4 Pro touch input
Root-cause fix: the AO3 library identified fics with std::hash<std::string>
(32-bit), but CrossInk's own epub cache migrated to ZipFile::fnvHash64
(64-bit) for its cache directory names a while back. Every AO3 sidecar
(ao3_library_info, holding the real file path plus all rich metadata) was
being written and read under the OLD hash scheme, but that no longer
matched where the shared epub cache actually put things once a book was
opened normally -- so on a from-scratch flash, metadata came back blank
(only the ~11-char index-fallback title showed) and tapping/selecting a fic
had no file path to open. xAO3 was unaffected because its CrossPoint base
predates the fnvHash64 migration. This was invisible until this session
because the AO3 port had never been exercised on real hardware.

Fix: switch fic-identity everywhere to ZipFile::fnvHash64 (matching
Epub::cachePathForFilePath), widening CompactIndexRecord/ViewEntry's
cacheHash from uint32_t to uint64_t. This changes the on-disk index record
size (241 -> 245 bytes) and hash values, so the index format version is
bumped 2 -> 3 (all read-side checks updated to match, including two in
Ao3LibraryActivity.cpp and Ao3EndOfBookSeriesActivity.cpp that were
initially missed and caused a false-negative "library index missing or
corrupt" on an otherwise-valid freshly-rebuilt index). An old v2 index is
rejected outright rather than misread; re-running Index New Books rebuilds
it cleanly.

Also fixes an indexing-loop bug found while rebuilding a large (500+ book)
library on-device: DIR_INDEXING only remembered the first 10 failed books
for re-processing exclusion, so once more than 10 books failed to scrape,
the excess were endlessly re-collected every batch and the progress
counter ran past the total without ever reaching completion. Failed
hashes are now tracked in a separate uncapped vector for exclusion;
failedBooks stays capped at 10 for the failed-list UI only.

Touch input for the X4 Pro (CROSSINK_APP_CAP_TOUCH), completing what
BookFusion already had and AO3 (ported button-only from xAO3) didn't:
Ao3LibraryActivity (library list tap-to-open/swipe-page/filter-tap, manage
panel, filter panel, fandom/relationship pickers), Ao3IndexActivity
(tap-to-confirm through the indexing prompts), Ao3LibrarySettingsActivity
(tap a row, cleanup prompts), Ao3FolderPickerActivity (tap to select,
long-press to navigate into a folder, matching the existing short/hold-
Confirm semantics). Every touch query used is a constexpr no-op when
CROSSINK_APP_CAP_TOUCH=0 (verified via the build's own
check_app_touch_gate.py audit), so default/sticky remain button-only.

Row hit-testing is done via TouchRegistry (registered at render time,
read back with wasItemTapped/wasScreenTapped), the same mechanism
BookFusion's screen.list relies on -- not hand-computed row geometry,
which proved unreliable on-device. In-place actions route through a
local per-frame flag rather than injectRelease(), which the main loop
only clears on the button-shortcut dispatch path, not the normal
activity loop -- an injected release there would never clear and would
fire every frame forever (this was the cause of the filter/manage panel
infinite-cycling bug found on-device).

Hardware-verified on X4 Pro: full library re-index (500+ books)
completes, metadata populates, tapping a fic opens it, filter/manage
panels and folder picker work via touch, buttons still work unchanged.

+++ src/Ao3ViewEntry.h
@@ -22,12 +22,12 @@ inline uint32_t fnv1a(const char* str) {
  * @brief In-RAM sort/filter key struct — one per live book, loaded sequentially
  *        from ao3_library_index.bin at library startup.
  *
- * 38 bytes packed (pragma pack 1).
- * 38 × 1000 books = 38 KB peak RAM.
+ * 42 bytes packed (pragma pack 1).
+ * 42 × 1000 books = 42 KB peak RAM.
  */
 #pragma pack(push, 1)
 struct ViewEntry {
-  uint32_t cacheHash;      // same as CompactIndexRecord.cacheHash
+  uint64_t cacheHash;      // same as CompactIndexRecord.cacheHash
   uint32_t wordCount;      // word count sort
   uint32_t seriesHash;     // fnv1a(seriesName), 0 if no series
   uint16_t addedSequence;  // date-added sort (monotonic, higher = newer), max 65535
